Move db layer to layers/db
